+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/PrometheusHttpListenerOptions.cs
@@ -18,6 +18,14 @@ public class PrometheusHttpListenerOptions
     private IReadOnlyCollection<string> uriPrefixes = ["http://localhost:9464/"];
 
     /// <summary>
+    /// Initializes a new instance of the <see cref="PrometheusHttpListenerOptions"/> class.
+    /// </summary>
+    public PrometheusHttpListenerOptions()
+    {
+        this.ScrapeResponseCacheDurationMilliseconds = 300;
+    }
+
+    /// <summary>
     /// Gets or sets the Host name the HTTP listener will bind to. Defaults to <c>localhost</c>.
     /// </summary>
     public string Host { get; set; } = "localhost";
@@ -36,6 +44,22 @@ public class PrometheusHttpListenerOptions
     /// Gets or sets a value indicating whether addition of _total suffix for counter metric names is disabled. Default value: <see langword="false"/>.
     /// </summary>
     public bool DisableTotalNameSuffixForCounters { get; set; }
+
+    /// <summary>
+    /// Gets or sets the cache duration in milliseconds for scrape responses. Default value: 300.
+    /// </summary>
+    /// <remarks>
+    /// Note: Specify 0 to disable response caching.
+    /// </remarks>
+    public int ScrapeResponseCacheDurationMilliseconds
+    {
+        get => field;
+        set
+        {
+            Guard.ThrowIfOutOfRange(value, min: 0);
+            field = value;
+        }
+    }
 
     /// <summary>
     /// Gets or sets the URI (Uniform Resource Identifier) prefixes to use for the http listener.